[WIP] remove NormalizedImports, start moving to Dependency enum on FileModule
