+++ src/commands/mod.rs
@@ -1,4 +1,5 @@
 pub mod check;
+pub mod helpers;
 pub mod report;
 pub mod server;
 pub mod sync;
