+++ src/processors/file_module.rs
@@ -1,23 +1,30 @@
-use std::marker::PhantomData;
-use std::path::PathBuf;
 use std::{path::Path, sync::Arc};
 
 use crate::filesystem::ProjectFile;
 use crate::{config::ModuleConfig, modules::ModuleNode};
 
 use super::dependency::Dependency;
-use super::import::{AllImports, ExternalImports, NormalizedImports, ProjectImports};
-
+use super::ignore_directive::IgnoreDirectives;
+use super::import::NormalizedImport;
+use super::reference::SourceCodeReference;
 #[derive(Debug)]
-pub struct FileModule<'a, State = AllImports> {
+pub struct FileModule<'a> {
     pub file: ProjectFile<'a>,
     pub module: Arc<ModuleNode>,
-    pub imports: NormalizedImports<State>,
+    pub ignore_directives: IgnoreDirectives,
     pub dependencies: Vec<Dependency<'a>>,
-    _state: PhantomData<State>,
 }
 
-impl<State> FileModule<'_, State> {
+impl<'a> FileModule<'a> {
+    pub fn new(file: ProjectFile<'a>, module: Arc<ModuleNode>) -> Self {
+        Self {
+            file,
+            module,
+            ignore_directives: IgnoreDirectives::empty(),
+            dependencies: vec![],
+        }
+    }
+
     pub fn module_config(&self) -> &ModuleConfig {
         self.module.config.as_ref().unwrap()
     }
@@ -25,85 +32,30 @@ impl<State> FileModule<'_, State> {
     pub fn relative_file_path(&self) -> &Path {
         &self.file.relative_file_path
     }
+
+    pub fn imports(&self) -> impl Iterator<Item = &NormalizedImport> {
+        self.dependencies.iter().filter_map(|dependency| {
+            if let Dependency::Import(import) = dependency {
+                Some(import)
+            } else {
+                None
+            }
+        })
+    }
+
+    pub fn references(&self) -> impl Iterator<Item = &SourceCodeReference> {
+        self.dependencies.iter().filter_map(|dependency| {
+            if let Dependency::Reference(reference) = dependency {
+                Some(reference)
+            } else {
+                None
+            }
+        })
+    }
 }
 
-impl<'a, State> AsRef<FileModule<'a, State>> for FileModule<'a, State> {
-    fn as_ref(&self) -> &FileModule<'a, State> {
+impl<'a> AsRef<FileModule<'a>> for FileModule<'a> {
+    fn as_ref(&self) -> &FileModule<'a> {
         self
-    }
-}
-
-impl<'a> FileModule<'a, AllImports> {
-    pub fn new(
-        file: ProjectFile<'a>,
-        module: Arc<ModuleNode>,
-        imports: NormalizedImports<AllImports>,
-        dependencies: Vec<Dependency<'a>>,
-    ) -> Self {
-        Self {
-            file,
-            module,
-            imports,
-            dependencies,
-            _state: PhantomData,
-        }
-    }
-
-    pub fn into_internal(self, source_roots: &[PathBuf]) -> FileModule<'a, ProjectImports> {
-        FileModule {
-            file: self.file,
-            module: self.module,
-            imports: self.imports.into_project_imports(source_roots),
-            dependencies: self.dependencies, // TODO: filter dependencies
-            _state: PhantomData,
-        }
-    }
-
-    pub fn into_external(self, source_roots: &[PathBuf]) -> FileModule<'a, ExternalImports> {
-        FileModule {
-            file: self.file,
-            module: self.module,
-            imports: self.imports.into_external_imports(source_roots),
-            dependencies: self.dependencies, // TODO: filter dependencies
-            _state: PhantomData,
-        }
-    }
-}
-
-pub type FileModuleInternal<'a> = FileModule<'a, ProjectImports>;
-
-impl<'a> FileModuleInternal<'a> {
-    pub fn new(
-        file: ProjectFile<'a>,
-        module: Arc<ModuleNode>,
-        imports: NormalizedImports<ProjectImports>,
-        dependencies: Vec<Dependency<'a>>,
-    ) -> Self {
-        Self {
-            file,
-            module,
-            imports,
-            dependencies,
-            _state: PhantomData,
-        }
-    }
-}
-
-pub type FileModuleExternal<'a> = FileModule<'a, ExternalImports>;
-
-impl<'a> FileModuleExternal<'a> {
-    pub fn new(
-        file: ProjectFile<'a>,
-        module: Arc<ModuleNode>,
-        imports: NormalizedImports<ExternalImports>,
-        dependencies: Vec<Dependency<'a>>,
-    ) -> Self {
-        Self {
-            file,
-            module,
-            imports,
-            dependencies,
-            _state: PhantomData,
-        }
     }
 }
