+++ src/processors/mod.rs
@@ -1,7 +1,8 @@
 pub mod dependency;
 pub mod file_module;
+pub mod ignore_directive;
 pub mod import;
 pub mod reference;
 
 pub use dependency::{Dependency, ExternalDependencyExtractor, InternalDependencyExtractor};
-pub use file_module::{FileModule, FileModuleExternal, FileModuleInternal};
+pub use file_module::FileModule;
